analysis/requirements.txt
flask==2.0.1
gunicorn==20.1.0
transformers==4.20.1
redis==4.0.2
PyJWT==2.4.0
torch==1.12.0
python-dotenv==0.19.0
werkzeug==2.0.3
numpy==1.21.6